src/requirements.txt
django==4.0
psycopg2_binary==2.9.2
djangorestframework==3.13.1
drf-spectacular==0.21.1
factory-boy==3.2.1